refactor: rename edit_tool to edit_tools and move message to messages
